implementing app naming feature
